test: cover remote cleanup state reset and waiter release
