+++ test/remote-session.test.ts
@@ -1,5 +1,5 @@
 import { afterEach, expect, test, vi } from 'vitest';
-import { Remote } from '../src/lib';
+import { Remote, Session } from '../src/lib';
 
 afterEach(() => {
   vi.restoreAllMocks();
@@ -118,3 +118,93 @@ test('remote encodes PlayReady session IDs in request URLs', async () => {
     vi.unstubAllGlobals();
   }
 });
+
+test.each(['close', 'remove'] as const)(
+  'remote %s clears native and wrapper state when cleanup fails',
+  async (operation) => {
+    const fetch = vi
+      .fn<typeof globalThis.fetch>()
+      .mockResolvedValueOnce(Response.json({ id: 'session' }));
+    vi.stubGlobal('fetch', fetch);
+    const engine = new Remote({ keySystem: 'com.widevine.alpha', baseUrl: 'https://cdm.test' });
+    const native = await engine.createSession();
+    const session = new Session('temporary', engine, native);
+    const closed = vi.fn();
+    const removed = vi.fn();
+    native.addEventListener('closed', closed);
+    native.addEventListener('removed', removed);
+    session.addEventListener('removed', removed);
+    const keyId = '00112233445566778899aabbccddeeff';
+    const key = 'ffeeddccbbaa99887766554433221100';
+    fetch.mockResolvedValueOnce(Response.json({ keys: { [keyId]: key } }));
+    await session.update(new Uint8Array([1]));
+    expect(native.keys.get(keyId)).toBe(key);
+    expect(session.keys.get(keyId)).toBe(key);
+    expect(session.keyStatuses.size).toBe(1);
+    const waiting = expect(session.waitForLicenseRequest()).rejects.toThrow('Session closed');
+    fetch.mockRejectedValueOnce(new Error('Server unavailable'));
+    await expect(session[operation]()).rejects.toThrow('Server unavailable');
+    await waiting;
+    expect(native.keys.size).toBe(0);
+    expect(native.keyStatuses.size).toBe(0);
+    expect(engine.sessions.size).toBe(0);
+    expect(session.keys.size).toBe(0);
+    expect(session.keyStatuses.size).toBe(0);
+    await expect(session.closed).resolves.toBe('closed-by-application');
+    expect(() => session.pause()).toThrow('Session closed');
+    await expect(native.update(new Uint8Array())).rejects.toThrow('Session closed');
+    // Failed remote cleanup is terminal locally; repeated cleanup must not reopen it.
+    await native[operation]();
+    await session.close();
+    await expect(session.remove()).rejects.toThrow('Session closed');
+    expect(closed).toHaveBeenCalledOnce();
+    expect(removed).not.toHaveBeenCalled();
+    expect(fetch).toHaveBeenCalledTimes(3);
+  },
+);
+
+test.each(['timeout', 'restart', 'server error'] as const)(
+  'remote cleanup releases waiters after %s',
+  async (failure) => {
+    const fetch = vi
+      .fn<typeof globalThis.fetch>()
+      .mockResolvedValueOnce(Response.json({ id: 'session' }));
+    vi.stubGlobal('fetch', fetch);
+    const engine = new Remote({
+      keySystem: 'com.widevine.alpha',
+      baseUrl: 'https://cdm.test',
+      requestTimeoutMs: 50,
+    });
+    const session = await engine.createSession();
+    const removed = vi.fn();
+    session.addEventListener('removed', removed);
+    const waiting = expect(session.waitForKeys()).rejects.toThrow('Session closed');
+    if (failure === 'timeout') {
+      vi.useFakeTimers();
+      fetch.mockImplementationOnce(
+        async (_url, init) =>
+          new Response(
+            new ReadableStream({
+              start(controller) {
+                init?.signal?.addEventListener('abort', () =>
+                  controller.error(init.signal?.reason),
+                );
+              },
+            }),
+          ),
+      );
+    } else {
+      fetch.mockResolvedValueOnce(
+        Response.json({ error: 'Cleanup failed' }, { status: failure === 'restart' ? 404 : 500 }),
+      );
+    }
+    const closing = expect(session.remove()).rejects.toThrow();
+    if (failure === 'timeout') await vi.advanceTimersByTimeAsync(50);
+    await closing;
+    await waiting;
+    expect(removed).not.toHaveBeenCalled();
+    expect(engine.sessions.size).toBe(0);
+    await session.close();
+    expect(fetch).toHaveBeenCalledTimes(2);
+  },
+);
